plget_plotter: add plget plotter, initially for histograms
Found forces to collect some of the peaces in one place to get
histograms. Should be modified soon to be more generic with correct
titles, speed and more. Also expect to add not only histograms but
automatic plots, that is actually implemented but spread through
many scripts .... Tested only on rx-lat output for now.

Signed-off-by: Ivan Khoronzhuk <ivan.khoronzhuk@linaro.org>